Move Cassandra initialization stuff to separate file
